% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litsence.R
\name{litsence}
\alias{litsence}
\title{Literature search using litsence API from NCBI}
\usage{
litsence(keywords, n = 30)
}
\arguments{
\item{keywords}{keyword}

\item{n}{the number of item}
}
\value{
a data frame
}
\description{
Literature search using litsence API from NCBI
}
\examples{
litsence("gut microbiota")
}
\author{
ZhonghuiGai
}
