% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Translate English to Chinese or the verse using trans function from translate-shell translator, https://github.com/soimort/translate-shell}
\usage{
translate(input)
}
\arguments{
\item{input}{The sentences to be translated}
}
\value{
The translated English or Chinese sentences
}
\description{
Translate English to Chinese or the verse using trans function from translate-shell translator, https://github.com/soimort/translate-shell
}
\examples{
translate("中文")
translate("English")
}
\author{
ZhonghuiGai
}
