% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litshow.R
\name{litshow}
\alias{litshow}
\title{Show the result of litsence}
\usage{
litshow(litsence, n = 1:30, size = 15, color = FALSE)
}
\arguments{
\item{litsence}{the result of litsence function}

\item{n}{display number}

\item{size}{the text size}
}
\value{
a plain text
}
\description{
Show the result of litsence
}
\examples{
res <- litsence(keywords = "gut microbiota")
litshow(res, 1:30)
}
\author{
ZhonghuiGai
}
